restructure zi driver
