+++ src/qibolab/instruments/zhinst/__init__.py
@@ -1,0 +1,4 @@
+from .executor import Zurich
+from .pulse import ZhPulse
+from .sweep import ProcessedSweeps, classify_sweepers
+from .util import acquire_channel_name, measure_channel_name
